Create tracksum.py
I know I've been assigned to create a script called tbsum.py, but I
think this a big enough departure from what I've been doing so far that
it warrants creating a new script file.

tracksum.py creates a histogram of events, but requires that the events
have a value over 450, and that they are local maxima spatially.